refactor: move command creation to plugin dir
